* Feature: IPv6 fully supported. This includes whois, range blocking, IPv6 city lookup in live traffic, country blocking and all other security functions. See www.wordfence.com/blog/ for more info. * Feature: New scanning routine examines the wp_options table for executable code based on a new infection we are seeing that is well hidden. * Improvement: Prevent Googlebot from being blocked if user has configured a banned URL and Google tries to crawl it. * Improvement: Improved detection for additional Google crawlers especially if an IP PTR resolves to a .googlebot.com domain. * Fix: Fixed bug with https:// URLs not allowed in country blocking. * Fix: Fixed typos.
git-svn-id: https://plugins.svn.wordpress.org/wordfence/tags/6.0.1@1164420 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ lib/wfConfig.php
@@ -34,6 +34,7 @@ class wfConfig {
 				"scansEnabled_plugins" => false,
 				"scansEnabled_malware" => false,
 				"scansEnabled_fileContents" => false,
+				"scansEnabled_database" => false,
 				"scansEnabled_posts" => false,
 				"scansEnabled_comments" => false,
 				"scansEnabled_passwds" => false,
@@ -121,6 +122,7 @@ class wfConfig {
 				"scansEnabled_plugins" => false,
 				"scansEnabled_malware" => true,
 				"scansEnabled_fileContents" => true,
+				"scansEnabled_database" => true,
 				"scansEnabled_posts" => true,
 				"scansEnabled_comments" => true,
 				"scansEnabled_passwds" => true,
@@ -208,6 +210,7 @@ class wfConfig {
 				"scansEnabled_plugins" => false,
 				"scansEnabled_malware" => true,
 				"scansEnabled_fileContents" => true,
+				"scansEnabled_database" => true,
 				"scansEnabled_posts" => true,
 				"scansEnabled_comments" => true,
 				"scansEnabled_passwds" => true,
@@ -295,6 +298,7 @@ class wfConfig {
 				"scansEnabled_plugins" => false,
 				"scansEnabled_malware" => true,
 				"scansEnabled_fileContents" => true,
+				"scansEnabled_database" => true,
 				"scansEnabled_posts" => true,
 				"scansEnabled_comments" => true,
 				"scansEnabled_passwds" => true,
@@ -382,6 +386,7 @@ class wfConfig {
 				"scansEnabled_plugins" => false,
 				"scansEnabled_malware" => true,
 				"scansEnabled_fileContents" => true,
+				"scansEnabled_database" => true,
 				"scansEnabled_posts" => true,
 				"scansEnabled_comments" => true,
 				"scansEnabled_passwds" => true,
@@ -789,8 +794,11 @@ class wfConfig {
 		return true;
 	}
 	public static function enableAutoUpdate(){
-		wfConfig::set('autoUpdate', '1');	
-		wp_schedule_event(time(), 'daily', 'wordfence_daily_autoUpdate');
+		wfConfig::set('autoUpdate', '1');
+		wp_clear_scheduled_hook('wordfence_daily_autoUpdate');
+		if (is_main_site()) {
+			wp_schedule_event(time(), 'daily', 'wordfence_daily_autoUpdate');
+		}
 	}
 	public static function disableAutoUpdate(){
 		wfConfig::set('autoUpdate', '0');	
@@ -857,11 +865,12 @@ Options -ExecCGI
 		$upload_dir = wp_upload_dir();
 		return $upload_dir['basedir'] . '/.htaccess';
 	}
-	
+
 	/**
 	 * Add/Merge .htaccess file in the uploads directory to prevent code execution.
 	 *
 	 * @return bool
+	 * @throws wfConfigException
 	 */
 	public static function disableCodeExecutionForUploads() {
 		$uploads_htaccess_file_path = self::_uploadsHtaccessFilePath();
@@ -880,27 +889,32 @@ Options -ExecCGI
 		}
 		return true;
 	}
-	
+
 	/**
 	 * Remove script execution protections for our the .htaccess file in the uploads directory.
 	 *
 	 * @return bool
+	 * @throws wfConfigException
 	 */
 	public static function removeCodeExecutionProtectionForUploads() {
 		$uploads_htaccess_file_path = self::_uploadsHtaccessFilePath();
 		if (file_exists($uploads_htaccess_file_path)) {
 			$htaccess_contents = file_get_contents($uploads_htaccess_file_path);
-			$htaccess_contents = str_replace(self::$_disable_scripts_htaccess, '', $htaccess_contents);
-			
-			$error_message = "Unable to remove code execution protections applied to the .htaccess file in the uploads directory.  Please check your permissions on that file.";
-			if (strlen(trim($htaccess_contents)) === 0) {
-				// empty file, remove it
-				if (!@unlink($uploads_htaccess_file_path)) {
+
+			// Check that it is in the file
+			if (strpos($htaccess_contents, self::$_disable_scripts_htaccess) !== false) {
+				$htaccess_contents = str_replace(self::$_disable_scripts_htaccess, '', $htaccess_contents);
+
+				$error_message = "Unable to remove code execution protections applied to the .htaccess file in the uploads directory.  Please check your permissions on that file.";
+				if (strlen(trim($htaccess_contents)) === 0) {
+					// empty file, remove it
+					if (!@unlink($uploads_htaccess_file_path)) {
+						throw new wfConfigException($error_message);
+					}
+
+				} elseif (@file_put_contents($uploads_htaccess_file_path, $htaccess_contents, LOCK_EX) === false) {
 					throw new wfConfigException($error_message);
 				}
-				
-			} elseif (@file_put_contents($uploads_htaccess_file_path, $htaccess_contents, LOCK_EX) === false) {
-				throw new wfConfigException($error_message);
 			}
 		}
 		return true;
